+++ src/main/kotlin/com/young_zy/forum/model/thread/SearchResultDTO.kt
@@ -1,11 +1,10 @@
 package com.young_zy.forum.model.thread
 
 import com.young_zy.forum.model.user.SimpleUserObject
-import java.math.BigInteger
 import java.time.LocalDateTime
 
 class SearchResultDTO(
-        var tid: BigInteger,
+        var tid: Long,
         var title: String,
         var lastReplyTime: LocalDateTime,
         var postTime: LocalDateTime,
